+++ fedify/federation/kv.ts
@@ -1,12 +1,12 @@
 /**
- * A key for a key-value store.  An array of one or more strings.
+ * A key for a key–value store.  An array of one or more strings.
  *
  * @since 0.5.0
  */
 export type KvKey = readonly [string] | readonly [string, ...string[]];
 
 /**
- * Additional options for setting a value in a key-value store.
+ * Additional options for setting a value in a key–value store.
  *
  * @since 0.5.0
  */
@@ -18,7 +18,7 @@ export interface KvStoreSetOptions {
 }
 
 /**
- * An abstract interface for a key-value store.
+ * An abstract interface for a key–value store.
  *
  * @since 0.5.0
  */
@@ -47,7 +47,7 @@ export interface KvStore {
 }
 
 /**
- * A key-value store that stores values in memory.
+ * A key–value store that stores values in memory.
  * Do not use this in production as it does not persist values.
  *
  * @since 0.5.0